split up index.js in multiple files
