+++ index.js
@@ -1,236 +1,15 @@
+'use strict';
+
 var assert = require('assert');
-var crypto = require('crypto');
-var moment = require('moment');
-var objectAssign = require('object-assign');
-var FormData = require('form-data');
-var stream = require('stream');
-var https = require('https');
-var request = require('request');
-var async = require('async');
-var fileType = require('file-type');
-
-var host = 'news-api.apple.com';
-var CRLF = '\r\n';
-
-var validContentTypes = [
-  'application/octet-stream', // should be first in this list
-  'image/jpeg',
-  'image/png',
-  'image/gif'
-];
-
-function encodeFormData (formData, cb) {
-  var form = new FormData();
-
-  async.mapSeries(formData, function (item, callback) {
-    var value = item.value;
-    var options = item.options;
-
-    function append () {
-      var len = typeof value === 'string' ? Buffer.byteLength(value) : value.length;
-      var header = '--' + form.getBoundary() + CRLF +
-        'Content-Type: ' + options.contentType + CRLF +
-        'Content-Disposition: form-data' +
-        (item.filename ? '; filename=' + encodeURIComponent(options.filename) : '') +
-        (item.name ? '; name=' + encodeURIComponent(item.name) : '') +
-        '; size=' + len + CRLF + CRLF;
-
-      form.append(item.name || item.filename, value,
-        objectAssign({ header: header, knownLength: len }, options));
-      callback();
-    }
-
-    if (options.contentType === 'application/json') {
-      return append();
-    }
-
-    request.get(value, {
-      timeout: 5000,
-      encoding: null,
-      rejectUnauthorized: process.env.NODE_ENV !== 'test'
-    }, function (err, res, body) {
-      if (err) {
-        return callback(err);
-      }
-
-      var contentType = fileType(body).mime;
-      if (validContentTypes.indexOf(contentType) === -1) {
-        contentType = validContentTypes[0];
-      }
-
-      options.contentType = contentType;
-      value = body;
-      return append();
-    });
-  }, function (err, callback) {
-    if (err) {
-      return cb(err);
-    }
-
-    var converter = new stream.Writable();
-    var chunks = [];
-    converter._write = function (chunk, enc, cb) {
-      chunks.push(chunk);
-      cb();
-    };
-
-    converter.on('finish', function () {
-      return cb(null, {
-        headers: form.getHeaders(),
-        buffer: Buffer.concat(chunks)
-      });
-    });
-
-    form.pipe(converter);
-  });
-}
+var setupMakeRequest = require('./lib/make-request');
+var createArticleUploadFormData = require('./lib/create-article-upload-form-data');
+var articleMetadataFromOpts = require('./lib/article-metadata-from-opts');
 
 module.exports = function (config) {
   assert(typeof config.apiId === 'string', 'config.apiId: API ID is required');
   assert(typeof config.apiSecret === 'string', 'config.apiSecret: API secret is required');
 
-  function sendRequest (method, endpoint, post, cb) {
-    var date = moment().utc().format('YYYY-MM-DDTHH:mm:ss[Z]');
-    var canonicalRequest = Buffer(method + 'https://' + (config.host || host) + endpoint + date +
-      (post ? post.headers['content-type'] : ''), 'ascii');
-    if (post) {
-      canonicalRequest = Buffer.concat([canonicalRequest, post.buffer]);
-    }
-
-    var key = new Buffer(config.apiSecret, 'base64');
-    var signature = crypto.createHmac('sha256', key)
-      .update(canonicalRequest, 'utf8')
-      .digest('base64');
-    var auth = 'HHMAC; key="' + config.apiId +
-      '"; signature="' + signature +
-      '"; date="' + date + '"';
-
-    var req = https.request({
-      method: method,
-      host: config.host || host,
-      port: config.port || void 0,
-      rejectUnauthorized: process.env.NODE_ENV !== 'test',
-      path: endpoint,
-      headers: objectAssign({
-        Accept: 'application/json',
-        Authorization: auth
-      }, post ? post.headers : {})
-    });
-
-    req.on('response', function (res) {
-      var result = '';
-      var done = false;
-
-      res.on('data', function (chunk) {
-        result += chunk.toString();
-      });
-
-      res.on('error', function (err) {
-        if (!done) {
-          done = true;
-          cb(err);
-        }
-      });
-
-      res.on('end', function () {
-        if (!done) {
-          done = true;
-          var parsed = null;
-
-          if (!result) {
-            return cb(null, res, null);
-          }
-
-          try {
-            parsed = JSON.parse(result);
-          } catch (e) {
-            return cb(e);
-          }
-
-          if (parsed.data) {
-            return cb(null, res, parsed.data);
-          }
-
-          if (parsed.errors && Array.isArray(parsed.errors) &&
-              parsed.errors.length > 0 && parsed.errors[0].code) {
-            var e = new Error(result);
-            e.apiError = parsed.errors[0];
-            return cb(e);
-          }
-
-          return cb(new Error(result));
-        }
-      });
-    });
-
-    if (post) {
-      req.write(post.buffer);
-    }
-
-    req.end();
-  }
-
-  function makeRequest (method, endpoint, requestOpts, cb) {
-    if (method === 'POST' && requestOpts.formData) {
-      return encodeFormData(requestOpts.formData || {}, function (err, encoded) {
-        if (err) {
-          return cb(err);
-        }
-
-        sendRequest(method, endpoint, encoded, cb);
-      });
-    }
-
-    sendRequest(method, endpoint, null, cb);
-  }
-
-  function createArticleUploadFormData (article, bundleFiles, metadata) {
-    bundleFiles = bundleFiles || {};
-    assert(typeof bundleFiles['article.json'] === 'undefined', 'bundle cannot contain article.json file');
-    assert(typeof bundleFiles['metadata'] === 'undefined', 'bundle cannot contain metadata file');
-    metadata = metadata || {};
-    var articleJson = JSON.stringify(article);
-    var result = [{
-      name: 'article.json',
-      filename: 'article.json',
-      value: articleJson,
-      options: {
-        filename: 'article.json',
-        contentType: 'application/json'
-      }
-    }, {
-      name: 'metadata',
-      value: JSON.stringify({ data: metadata }),
-      options: {
-        filename: 'metadata',
-        contentType: 'application/json'
-      }
-    }];
-
-    Object.keys(bundleFiles).forEach(function (name, index) {
-      var url = bundleFiles[name];
-      result.push({
-        filename: name,
-        name: 'file' + index,
-        value: url,
-        options: {
-          filename: name
-        }
-      });
-    });
-
-    return result;
-  }
-
-  function articleMetadataFromOpts (opts) {
-    assert(typeof opts.isPreview === 'undefined' || typeof opts.isPreview === 'boolean');
-    assert(typeof opts.isSponsored === 'undefined' || typeof opts.isSponsored === 'boolean');
-
-    return {
-      isPreview: typeof opts.isPreview === 'boolean' ? opts.isPreview : true,
-      isSponsored: !!opts.isSponsored
-    };
-  }
+  var makeRequest = setupMakeRequest(config);
 
   return {
     readChannel: function (opts, cb) {
