+++ lib/article-metadata-from-opts.js
@@ -1,0 +1,13 @@
+'use strict';
+
+var assert = require('assert');
+
+module.exports = function articleMetadataFromOpts (opts) {
+  assert(typeof opts.isPreview === 'undefined' || typeof opts.isPreview === 'boolean');
+  assert(typeof opts.isSponsored === 'undefined' || typeof opts.isSponsored === 'boolean');
+
+  return {
+    isPreview: typeof opts.isPreview === 'boolean' ? opts.isPreview : true,
+    isSponsored: !!opts.isSponsored
+  };
+};
